requirements.txt
langchain
langchain-community
langchain-openai
python-dotenv
pypdf
chromadb
fastapi
uvicorn
langchain-google-genai